+++ src/adapter/bun/handler-native.ts
@@ -39,6 +39,6 @@ export const createNativeStaticHandler = (
 		if (!response.headers.has('content-type'))
 			response.headers.append('content-type', 'text/plain')
 
-		return () => response.clone() as any
+		return () => response.clone() as Response
 	}
 }