+++ src/adapter/bun/handler.ts
@@ -664,7 +664,7 @@ export const createStaticHandler = (
 		!hooks.beforeHandle?.length &&
 		!hooks.afterHandle?.length
 	)
-		return response.clone.bind(response) as any
+		return () => response.clone() as Response
 }
 
 const handleResponse = createResponseHandler({